[clean] Warning correction + clean code

+++ BDE/BDE/Member.h
@@ -6,11 +6,10 @@ class Member : public Student
 protected:
 	float reduction;
 public:
-	Member(string name, string firstName, int reduction) : Student(name, firstName), reduction(reduction) { ; }
+	Member(string name, string firstName, float reduction) : Student(name, firstName), reduction(reduction) { ; }
 	
 	
 	virtual float getPriceOfParty(float brutePrice) {
-		//cout << "Hello I am a Debugger";
 		float newPrice = brutePrice * (1 - (reduction / 100));
 		return newPrice;
 	}
